move all service methods into interface definition
